Refactor mapping process
